[FIX] slot: nested t-slot multiple levels
When a Component was nested in a slot within another directly
that is, in the same template, its rendering context was wrongly
its *parent*, whereas it should be the component which we *are* rendering
the template of.

A similar issue happened when a t-slot was nested within a slot, itself
having a t-slot.
In that case though, the crash in a browser was a memory error

This commit fixes both issues by introducing the concept of contextualParent
which is the Component of which we are compiling the template

At runtime (after compile time) this contextualParent is irrelevant
it doesn't take any part in the life cycle of any component

closes #545