[IMP] reserving __target key to return target of Observer's proxy
The return of useState is basically a Proxy returned by observe of
Observer instance. I'm not fully sure but there might be instances that
we want the real object consumed by useState. If for example we want to
compare the state's original obj to another obj that might actually be
the same to the original, then perhaps having access to the original obj
can be important. We can actually skip shallow or deep compares here as
we know that the returning obj is just the original obj.

With this commit, I'm suggesting to reserve the `__target` key such that
if accessed thru the state, it returns the original obj consumed by
useState.

Sample usage:

--in the constructor--
this.state = useState(this.props.originalObj);

--somewhere inside one of component's methods --
const originalObj = this.state.__target