[FIX] component: properly clean child vnode on destroy
When a child component is destroyed, the method does not clean its vnode from
its parent's children list, because it is assumed for performance that the
parent will be destroyed as well.

However, if the parent is not destroyed, it will be left with obsolete elements
in its children list. When that is the case, if a following render is bringing
back a previously deleted child, this child will not be added back in the DOM
due to the parent thinking it is already there.