[FIX] component: allow var with body as (textual) props
Before this commit, a variable set in a template with t-set could be
passed to a subcomponent as a prop only if it was only textual. If the
value was defined in the body of the t-set directive, it was passed, but
as a VDomArray, which means that it will only be displayed as [object
object] or something.

Also, we do not want to leak the VDomArray abstraction in userland (code
written by people using owl).

This issue is actually quite a problem in practice, because values in a
templates are translated, but not in attributes.  Therefore, using a
t-set directive with a body text content is the proper way to have
translated values at runtime.

closes #670